refacto: one page container component
