+++ web/landing/accessibility.js
@@ -1,39 +1,21 @@
 import React from "react";
-import Container from "@mui/material/Container";
-import { Header } from "../common/Header";
-import { makeStyles } from "@mui/styles";
-import { Footer } from "./footer";
 import Link from "@mui/material/Link";
 import { usePageTitle } from "../common/UsePageTitle";
-import { Main } from "../common/semantics/Main";
-
-const useStyles = makeStyles(theme => ({
-  container: {
-    padding: theme.spacing(4),
-    paddingBottom: theme.spacing(7),
-    margin: "auto",
-    textAlign: "left"
-  }
-}));
+import { FullPageComponent } from "./components/FullPageComponent";
 
 export default function Accessibility() {
   return (
-    <>
-      <Header />
-      <Main>
-        <AccessibilityDeclaration />
-      </Main>
-      <Footer />
-    </>
+    <FullPageComponent>
+      <AccessibilityDeclaration />
+    </FullPageComponent>
   );
 }
 
 function AccessibilityDeclaration() {
   usePageTitle("Déclaration d'accessibilité - Mobilic");
-  const classes = useStyles();
 
   return (
-    <Container className={classes.container} maxWidth="xl">
+    <>
       <h1>Déclaration d’accessibilité</h1>
       <p>
         Établie le <span>27 mars 2023</span>.
@@ -132,6 +114,6 @@ function AccessibilityDeclaration() {
         </Link>
         .
       </p>
-    </Container>
+    </>
   );
 }
